Create Compute_Grade.py
This is my first program using the concept of functions.

A function allows us to reuse a snippet of code multiple times within a program, following the DRY (Don't Repeat Yourself) principle. 

The return statement allows a function to send back a value, which can be used elsewhere in the program.

To pass a string value, we enclose it in single (') or double (") quotes. 

Parameters inside a function are local variables, meaning they exist only within that function and do not interact with outer variables unless explicitly returned or modified.